refactor: hamburger menu refactor
