--- frontend/src/components/Hamburger/MenuContent/MenuContent.js
@@ -1,34 +1,0 @@
-import React from "react";
-import PropTypes from "prop-types";
-import { Link } from "react-router-dom";
-import { Wrapper, ListWrapper, List, ListItems, NickNameDisplayWrapper, NickName } from "./style";
-
-const HamburgerMenuContent = (props) => {
-  return (
-    <Wrapper>
-      <NickNameDisplayWrapper>
-        <NickName>닉네임</NickName>
-      </NickNameDisplayWrapper>
-      <ListWrapper>
-        <List>
-          <ListItems>
-            <a>내가 쓴 글</a>
-          </ListItems>
-          <ListItems>
-            <a>내 정보 수정</a>
-          </ListItems>
-          <ListItems>
-            <Link to="/write/1">글 작성</Link>
-          </ListItems>
-          <ListItems>
-            <a>로그아웃</a>
-          </ListItems>
-        </List>
-      </ListWrapper>
-    </Wrapper>
-  );
-};
-
-HamburgerMenuContent.propTypes = {};
-
-export default HamburgerMenuContent;
