--- frontend/src/components/Hamburger/index.js
@@ -1,4 +1,0 @@
-import HamburgerMenu from "./MenuButton/MenuButton";
-import HamburgerModal from "./Modal/Modal";
-
-export { HamburgerMenu, HamburgerModal };
