+++ frontend/src/components/HamburgerMenu/index.js
@@ -1,0 +1,4 @@
+import HamburgerMenu from "./MenuButton";
+import HamburgerModal from "./Modal";
+
+export { HamburgerMenu, HamburgerModal };
